Revert "LLM: bigdl-llm lora support & lora example (#9740)"
This reverts commit 48342409698fda77a848554ff0c08ce86de67faa.
