+++ python/llm/example/GPU/QLoRA-FineTuning/alpaca-qlora/README.md
@@ -1,6 +1,6 @@
-# Alpaca Finetuning with BigDL-LLM
+# Alpaca QLoRA & QA-LoRA Finetuning (experimental support)
 
-This example ports [Alpaca-LoRA](https://github.com/tloen/alpaca-lora/tree/main) to BigDL-LLM (using either [QLoRA](https://arxiv.org/abs/2305.14314) / [QA-LoRA](https://arxiv.org/abs/2309.14717) or [LoRA](https://arxiv.org/abs/2106.09685) algorithm) on [Intel GPU](../../README.md).
+This example ports [Alpaca-LoRA](https://github.com/tloen/alpaca-lora/tree/main) to BigDL-LLM (using either [QLoRA](https://arxiv.org/abs/2305.14314) or [QA-LoRA](https://arxiv.org/abs/2309.14717) algorithm) on [Intel GPU](../../README.md).
 
 ### 0. Requirements
 To run this example with BigDL-LLM on Intel GPUs, we have some recommended requirements for your machine, please refer to [here](../../README.md#requirements) for more information.
@@ -25,8 +25,6 @@ source /opt/intel/oneapi/setvars.sh
 ```
 
 ### 3. Finetune
-
-Now we support three training modes ([QLoRA](https://arxiv.org/abs/2305.14314) / [QA-LoRA](https://arxiv.org/abs/2309.14717) / [LoRA](https://arxiv.org/abs/2106.09685)), to run different mode, just change `training_mode` to `qlora` / `qalora` / `lora` in below script.
 
 Here, we provide example usages on different hardware. Please refer to the appropriate script based on your device:
 
@@ -97,26 +95,6 @@ bash qalora_finetune_llama2_7b_arc_2_card.sh
 
 ```bash
 bash qalora_finetune_llama2_7b_pvc_1550_1_tile.sh
-```
-
-#### LoRA
-
-##### Finetuning LLaMA2-7B on four Intel Data Center GPU Max 1100
-
-```bash
-bash lora_finetune_llama2_7b_pvc_1100_1_card.sh
-```
-
-##### Finetuning LLaMA2-7B on single Tile Intel Data Center GPU Max 1550
-
-```bash
-bash lora_finetune_llama2_7b_pvc_1550_1_tile.sh
-```
-
-##### Finetuning LLaMA2-7B on four Intel Data Center GPU Max 1550
-
-```bash
-bash lora_finetune_llama2_7b_pvc_1550_4_card.sh
 ```
 
 ### 4. (Optional) Resume Training
