+++ python/llm/example/GPU/QLoRA-FineTuning/alpaca-qlora/qalora_finetune_llama2_7b_arc_2_card.sh
@@ -31,4 +31,4 @@ mpirun -n 2 \
        --lora_alpha 16 \
        --lora_dropout 0.05 \
        --val_set_size 2000 \
-    --training_mode "qalora" > training.log
+       --qa_lora True > training.log
