+++ python/llm/example/GPU/QLoRA-FineTuning/alpaca-qlora/qalora_finetune_llama2_7b_pvc_1550_1_card.sh
@@ -24,7 +24,7 @@ mpirun -n 2 \
        --base_model "meta-llama/Llama-2-7b-hf" \
        --data_path "yahma/alpaca-cleaned" \
        --output_dir "./bigdl-qlora-alpaca" \
-       --training_mode "qalora" \
+       --qa_lora True \
        --learning_rate 9e-5 \
        --micro_batch_size 8 \
        --batch_size 128 \
